requirements.txt
streamlit==1.37.0
pyyaml==6.0.2
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
altair==5.3.0
